Add a method to select random data from a range
This range is given by a SQL Query, and the top will be
the count() number returned by the SQL Query.

The logic behind this is that I need two random indexes, both
must exist in the table Estudiante and Materia respectively,
so I created a SQL Query that returns the number of records in a
given table, then I just use a random property that chooses a
number between 0 - N, where N is the number returned by the SQL Query.